+++ tests/TestCase.php
@@ -76,8 +76,6 @@ GRAPHQL;
         /** @var \Illuminate\Contracts\Config\Repository $config */
         $config = $app['config'];
 
-        $config->set('lighthouse.route.prefix', 'test-api');
-
         $config->set('lighthouse.namespaces', [
             'models' => [
                 'Tests\\Utils\\Models',